Moves Pause function into HandleScrape file (cleanliness)
